Rozbudowa ModifyOperationWindow

+++ src/BackEnd/OperationDir/PlannedExpanse.java
@@ -3,7 +3,7 @@ package BackEnd.OperationDir;
 import java.util.Date;
 
 public class PlannedExpanse extends BackEnd.OperationDir.CyclicOperation {
-    public PlannedExpanse(double value, double balance, long ID, Date day, String tag, String description) {
+    public PlannedExpanse(double value, double balance, int ID, Date day, String tag, String description) {
         super(value, balance, ID, day, tag, description);
     }
 }
